Add form to submit user name
The browser will make a post request to the /greetings path (defined
inside the action attribute) passing the inserted data as fname and
lname (the values given to the name attribute for each input). Those
values will be stored in the req.body object as key/value pairs, where
the keys are the name attributes and the values are the inputed data.