Create a route using wildcards
I have defined a route using wildcards for the welcome page. The route
takes two wildcards. The values passed for each wildcard are stored as
keys in the req.params object, where the name of the wildcard in the
route is the key and the value passed to it is the value. For example, a
request made (by the browser) to /welcome/Matheus/Alves will result in
req.params.fname being equals to Matheus, and req.params.lname being
equals to Alves.